Performance enhancement.
NOTE: If you're using this and want to see the performance improvement you will need to restart the restore OR retract
the unique constraint on the ::cloning/original-id attribute in a transaction whose tx time is AFTER the last restored
transaction, but BEFORE the next one you need to restore. Incrementing the last transaction time by a millisecond for
this is your best bet.

* Made the original ID NOT unique (identified as a severe performance problem)
* Fixed some spec problems